[utils] Merge unic-utils-codepoints src into unic-utils
Closes https://github.com/behnam/rust-unic/issues/50
