+++ unic/ucd/age/Cargo.toml
@@ -18,4 +18,4 @@ travis-ci = { repository = "behnam/rust-unic", branch = "master" }
 unic-ucd-core = { path = "../core/", version = "0.4.0" }
 
 [dev-dependencies]
-unic-utils-codepoints = { path = "../../utils/codepoints", version = "0.4.0" }
+unic-utils = { path = "../../utils", version = "0.4.0" }
