--- unic/utils/codepoints/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "unic-utils-codepoints"
-version = "0.4.0"
-authors = ["The UNIC Project Developers"]
-homepage = "https://github.com/behnam/rust-unic/"
-repository = "https://github.com/behnam/rust-unic/"
-license = "MIT/Apache-2.0"
-keywords = ["text", "unicode"]
-description = "UNIC - Utilities for working with Unicode Code Points"
-
-# No tests/benches that depends on /data/
-exclude = []
-
-[badges]
-travis-ci = { repository = "behnam/rust-unic", branch = "master" }
