+++ unic/utils/src/codepoints.rs
@@ -9,8 +9,6 @@
 // except according to those terms.
 
 
-#![forbid(unsafe_code, missing_docs)]
-
 //! # UNIC — Utils — Code Points
 //!
 //! A component of [`unic`: Unicode and Internationalization Crates for Rust](/unic/).
@@ -22,18 +20,28 @@
 //! * [**Unicode Code Point**](http://unicode.org/glossary/#code_point)
 //! * [**Unicode Scalar Value**](http://unicode.org/glossary/#unicode_scalar_value)
 
-
-/// UNIC component version.
-pub const PKG_VERSION: &'static str = env!("CARGO_PKG_VERSION");
-
-/// UNIC component name.
-pub const PKG_NAME: &'static str = env!("CARGO_PKG_NAME");
-
-/// UNIC component description.
-pub const PKG_DESCRIPTION: &'static str = env!("CARGO_PKG_DESCRIPTION");
+use std::char::from_u32;
+use std::ops::Range;
 
 
-mod codepoints;
+/// Range of Unicode Code Points.
+///
+/// Reference: <http://unicode.org/glossary/#code_point>
+pub const CODEPOINTS_RANGE: Range<u32> = 0x0..(0x10FFFF + 1);
 
+/// Range of Unicde Scalar Values.
+///
+/// Reference: <http://unicode.org/glossary/#unicode_scalar_value>
+const SCALAR_VALUE_RANGE_1: Range<u32> = 0x0..0xD800;
+const SCALAR_VALUE_RANGE_2: Range<u32> = (0xDFFF + 1)..(0x10_FFFF + 1);
 
-pub use codepoints::{iter_all_chars, CODEPOINTS_RANGE};
+/// Check a code-point against `SURROGATE_CODEPOINTS_RANGE`.
+#[inline]
+pub fn iter_all_chars() -> Box<Iterator<Item = char>> {
+    Box::new(
+        // TODO: maybe use char::from_u32_unchecked()
+        SCALAR_VALUE_RANGE_1
+            .chain(SCALAR_VALUE_RANGE_2)
+            .filter_map(from_u32),
+    )
+}
