+++ unic/utils/src/lib.rs
@@ -26,7 +26,7 @@ pub const PKG_NAME: &'static str = env!("CARGO_PKG_NAME");
 pub const PKG_DESCRIPTION: &'static str = env!("CARGO_PKG_DESCRIPTION");
 
 
-pub extern crate unic_utils_codepoints as codepoints;
+pub mod codepoints;
 
 
 pub use codepoints::iter_all_chars;
